refactor: use getopt instead of switch-case

+++ dunst-snooze.sh
@@ -16,7 +16,7 @@ function log {
 }
 
 function pause_for_time {
-	if [[ "$CURRENT_VALUE" < 1 ]]; then
+	if [ "$CURRENT_VALUE" -lt 1 ]; then
 		dunstctl set-paused true &&
 			bash -c "$1" &
 	else
@@ -29,39 +29,74 @@ function pause_for_time {
 }
 
 function format_timer {
-	[[ "$1" < 1 ]] && printf "♾" && return
+	[ "$1" -lt 1 ] && printf "♾" && return
 	local hours=$(expr "$1" / 60)
 	local minutes=$(expr "$1" % 60)
 
-	[[ "$hours" < 1 ]] &&
+	[ "$hours" -lt 1 ] &&
 		printf "%02.f" "$minutes" ||
 		printf "%b:%02.f" "$hours" "$minutes"
 }
 
-case $1 in
-"--up")
-	CURRENT_VALUE=$(echo "$(($CURRENT_VALUE + 5))")
-	;;
-"--down")
-	[[ "$CURRENT_VALUE" > 0 ]] &&
-		CURRENT_VALUE=$(echo "$(($CURRENT_VALUE - 5))")
-	;;
-"--toggle")
-	[ -e "$PID_FILE" ] && kill -9 "$(cat $PID_FILE)"
-	log "pid file: $PID_FILE and pid is $$"
-	echo "$$" >"$PID_FILE"
+parse_command_line() {
+	ARGS=$(getopt \
+		-o t:ud \
+		--long toggle:,up,down \
+		-- "$@")
+	getopt_exit="$?"
 
-	[ "$(dunstctl is-paused)" = "true" ] &&
-		dunstctl set-paused false ||
-		pause_for_time "$2"
+	if [ "$getopt_exit" -ne 0 ]; then
+		log "don't know what you wanted to do"
+		exit 1
+	fi
 
-	bash -c "$2"
-	rm "$PID_FILE"
-	;;
-esac
+	eval set -- "$ARGS"
 
-echo "$CURRENT_VALUE" >"$TIMER_FILE"
+	for o; do
+		case "$o" in
+		-t | --toggle)
+			if [ -n "${2:-}" ]; then
+				local post_call="${2}"
+				[ -e "$PID_FILE" ] && kill -9 "$(cat $PID_FILE)"
+				log "pid file: $PID_FILE and pid is $$"
+				echo "$$" >"$PID_FILE"
 
-log "$TIMER_FILE - had value: $PERVIOUS_VALUE, now has: $CURRENT_VALUE"
+				[ "$(dunstctl is-paused)" = "true" ] &&
+					dunstctl set-paused false ||
+					pause_for_time "$post_call"
 
-echo "$(is_paused) $(format_timer "$CURRENT_VALUE")"
+				bash -c "$post_call"
+				rm "$PID_FILE"
+				shift 2
+			else
+				shift 1
+			fi
+			;;
+		-d | --down)
+			[ "$CURRENT_VALUE" -gt 0 ] &&
+				CURRENT_VALUE=$(echo "$(($CURRENT_VALUE - 5))")
+			shift 1
+			;;
+		-u | --up)
+			CURRENT_VALUE=$(echo "$(($CURRENT_VALUE + 5))")
+			shift 1
+			;;
+		--)
+			shift
+			break
+			;;
+		esac
+	done
+}
+
+function main {
+	parse_command_line "$@"
+
+	echo "$CURRENT_VALUE" >"$TIMER_FILE"
+
+	log "$TIMER_FILE - had value: $PERVIOUS_VALUE, now has: $CURRENT_VALUE"
+
+	echo "$(is_paused) $(format_timer "$CURRENT_VALUE")"
+}
+
+main "$@"
